+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # gqlyzer
 
+<a name="v1.3.2"></a>
+## [v1.3.2] - 2025-12-16
+### Fixes
+- ignore errors
+- fix parse query with text and space as separator
+
+
 <a name="v1.3.1"></a>
 ## [v1.3.1] - 2025-12-02
 ### Fixes
@@ -24,7 +31,8 @@
 <a name="v1.0.0"></a>
 ## v1.0.0 - 2019-09-13
 
-[Unreleased]: https://github.com/kumparan/gqlyzer/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/kumparan/gqlyzer/compare/v1.3.2...HEAD
+[v1.3.2]: https://github.com/kumparan/gqlyzer/compare/v1.3.1...v1.3.2
 [v1.3.1]: https://github.com/kumparan/gqlyzer/compare/v1.3.0...v1.3.1
 [v1.3.0]: https://github.com/kumparan/gqlyzer/compare/v1.2.0...v1.3.0
 [v1.2.0]: https://github.com/kumparan/gqlyzer/compare/v1.1.0...v1.2.0